services/tts/requirements.txt
fastapi
uvicorn[standard]
httpx>=0.28.1
python-dotenv